Fix bug that caused entered key to be eaten
When a enemy laser has a character that matches the first character of
a given word, that character was being eaten by the laser interception
code. If an interception laser has been fired already, the character
input should be allowed to pass through to be processed by as a part of
a word.