Introduce three different difficulty levels
On the Easy difficulty, only alphabet letters are used.

On the Normal difficulty, regular words are used.

On the Hard difficulty, it is the same as the Normal difficulty except
that the enemy ships will fire one laser at the player that needs to be
cleared by typing in a letter.