Create a WordManager for handling the sets of words
Abstracted out all of the logic for handling the sets of words to
present to the user into the new WordManager class. The WordManager
will be responsible for finding and presenting the next word to show
to the user as well as managing the different set of words that will
be shown to the user. The WordManager internally knows which set of
word the user is currently on and as such has the information required
for identifying whether the user has completed all the words or not.