Create levels in the game
Implemented level support into the game. After all the words have been
cleared, the user will again be prompted with the same words but the
speed in which the ships move will increase as the user advances
further in the game. The game will end after the user has completed
five levels.