Add a shield power up that enemies may drop
Enemies now have a chance of dropping a shield power up. Shields are
automatically picked up by the player. A shield protects the user from
a maximum of three hits. If the shield takes three hits, the shield
will be destroyed. If the player picks up a shield power up while the
player already has a shield, the shield will be restored to its maximm
health (if applicable).