Replace levels with a boss instead
Overhauled the code base to create a generic abstract concept of a
stage. After the player has completed all of the words, the player
will battle a boss instead. Upon defeating the boss, the player will
win the game.

Also implemented shift key support to add support for words with
capital letters such as the days of the week and the months of the
year.