Implement custom word list support
Create a function to allow the player to open text files containing a
word list and to then use those words in the game. The word list should
be in the same format as the files that are used by Zodiac. All opened
files will be collapsed into a single flat list of words.